refactor: reservation service logic
