+++ src/reservation/dto/request-dto/reservation-param.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsDateString, IsNotEmpty } from 'class-validator';
 
-export class ReservationQueryDto {
+export class ReservationParamDto {
   @ApiProperty({
     description: '조회를 원하는 예약 날짜',
     example: '2022-12-12',
