+++ src/reservation/entities/reservation.entity.ts
@@ -1,7 +1,7 @@
 import { Hospital } from '../../hospital/entities/hospital.entity';
 import { Patient } from '../../patient/entities/patient.entity';
 import { User } from '../../user/entities/user.entity';
-import { Entity, Column, ManyToOne } from 'typeorm';
+import { Entity, Column, ManyToOne, JoinColumn } from 'typeorm';
 import { Common } from '../../entities/common.entity';
 
 @Entity()
@@ -21,13 +21,16 @@ export class Reservation extends Common {
   @ManyToOne(() => Hospital, (hospital) => hospital.reservations, {
     onDelete: 'CASCADE',
   })
+  @JoinColumn()
   hospital: Hospital;
 
   @ManyToOne(() => User, (user) => user.reservations, { onDelete: 'CASCADE' })
+  @JoinColumn()
   user: User;
 
   @ManyToOne(() => Patient, (patient) => patient.reservations, {
     onDelete: 'CASCADE',
   })
+  @JoinColumn()
   patient: Patient;
 }
